adds preview manager
