+++ Connectors/Rhino/Speckle.Connectors.RhinoShared/Speckle.Connectors.RhinoShared.projitems
@@ -21,6 +21,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Bindings\RhinoReceiveBinding.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Bindings\RhinoSendBinding.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Bindings\RhinoSelectionBinding.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)HostApp\RhinoPreviewManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Receive\RhinoMultiplayerHostObjectBuilder.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Operations\Send\Settings\EnableLiveSession.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Registration\ServiceRegistration.cs" />
